Move Client into sprockets_amqp.client.
