+++ docs/api.rst
@@ -3,6 +3,9 @@ API
 .. automodule:: sprockets.mixins.amqp
    :members:
 
+.. automodule:: sprockets_amqp.client
+   :members:
+
 Exceptions
 ----------
 .. automodule:: sprockets_amqp.exceptions
